EventTimeline : expose only one interface
